infra-api/requirements.txt
Django==4.2.4
requests==2.25.1
djangorestframework==3.14.0
Jinja2==3.1.2
pyodbc==5.0.1
werkzeug==3.0.1
django-extensions==3.2.3
pyOpenSSL==23.3.0
django-cors-headers==4.3.1
ldap3==2.9.1
PyFortiAPI==0.3.0
redis
celery
django-celery-results
elasticsearch
django-redis==5.4.0
pika==1.3.2
python-terraform==0.10.1
